% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colocPropTest-package.R
\docType{package}
\name{colocPropTest-package}
\alias{colocPropTest-package}
\alias{_PACKAGE}
\title{The colocPropTest package}
\description{
Conducts proportional colocalisation tests. See the main function
run_proptests().
}
\author{
\strong{Maintainer}: Chris Wallace \email{chris.x.wallace@gsk.com} (\href{https://orcid.org/0000-0001-9755-1703}{ORCID})

}
\keyword{internal}
