% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proportional.R
\name{estprop}
\alias{estprop}
\title{Proportional colocalisation testing supplying only a pair of regression
coefficients.}
\usage{
estprop(b1, b2, V1, V2)
}
\arguments{
\item{b1}{regression coefficients for trait 1, expect length(b1)=2}

\item{b2}{regression coefficients for trait 2, expect length(b2)=2}

\item{V1}{2x2 variance-covariance matrix for trait 1}

\item{V2}{2x2 variance-covariance matrix for trait 2}
}
\value{
a list, containing
* result: the test statistic
* plot.data: dataset for plotting the input data
* plot.eta: dataset for plotting chisq as a function of theta or eta
}
\description{
Proportional colocalisation testing supplying only a pair of regression
coefficients.
}
\author{
Chris Wallace
}
