% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proportional.R
\name{estprop_slow}
\alias{estprop_slow}
\title{Proportional colocalisation testing}
\usage{
estprop_slow(b1, b2, V1, V2)
}
\arguments{
\item{b1}{regression coefficients for trait 1}

\item{b2}{regression coefficients for trait 2}

\item{V1}{variance-covariance matrix for trait 1}

\item{V2}{variance-covariance matrix for trait 2}
}
\value{
a list, containing the test statistic and two datasets for
    plotting the input data or eta
}
\description{
This should return the same as estprop for a pair of snps, but is
slower. Left here for checking. Also accomodates more than two
snps.
}
\author{
Chris Wallace
}
