% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{keep_from_S}
\alias{keep_from_S}
\title{keep snp subset of coloc dataset}
\usage{
keep_from_S(S, keep)
}
\arguments{
\item{S}{coloc dataset}

\item{keep}{snps to keep}
}
\value{
subset of coloc dataset
}
\description{
keep snp subset of coloc dataset
}
