% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{lp}
\alias{lp}
\title{logp}
\usage{
lp(beta, se)
}
\arguments{
\item{beta}{coefficient}

\item{se}{std error of coefficient}
}
\value{
-log10 p
}
\description{
uses logs in calculation to avoid numerical issues with very small std errors / p values
}
