% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proportional.R
\name{marg_with_V}
\alias{marg_with_V}
\title{create variance-covariance matrix for pair of marginal beta + vbeta, given
estimate of r between snps}
\usage{
marg_with_V(beta, vbeta, rho)
}
\arguments{
\item{beta}{vector of two coefficients at two snps}

\item{vbeta}{vector of two variance of coefficients at the same two snps}

\item{rho}{LD (r) between the two snps}
}
\value{
list of coefficient & variance-covariance matrix
}
\description{
create variance-covariance matrix for pair of marginal beta + vbeta, given
estimate of r between snps
}
