% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tag}
\alias{tag}
\title{Derive tag SNPs using heirarchical clustering}
\usage{
tag(r2, r2_threshold = 0.95, quiet = FALSE, method = "complete")
}
\arguments{
\item{r2}{matrix of rsquared values}

\item{r2_threshold}{only 1 of a set of snps with r2 > r2_threshold will be kept}

\item{quiet}{if FALSE (default), show progress messages}

\item{method}{method used for heirarchical clustering.  See hclust for options.}
}
\value{
character vector, names are \code{snps}, values are the tag for each SNP
}
\description{
Uses complete linkage and the \code{\link{hclust}} function to define clusters,
then cuts the tree at 1-tag.threshold
}
\author{
Chris Wallace
}
