% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proportional.R
\name{tester_marg}
\alias{tester_marg}
\title{run proportional test directly on marginal test stats from coloc datasets}
\usage{
tester_marg(j, S1, S2, LD1, LD2 = LD1)
}
\arguments{
\item{j}{indices of thw two snps}

\item{S1}{coloc dataset 1}

\item{S2}{coloc dataset 2}

\item{LD1}{LD matrix for dataset 1 - rownames, colnames capture the snps and S1$snp[j] must
be represented}

\item{LD2}{LD matrix for dataset 2 - rownames, colnames capture the snps and S2$snp[j] must
be represented. if not supplied, defaults to LD1}
}
\value{
result from estprop
}
\description{
run proportional test directly on marginal test stats from coloc datasets
}
